+++ native/src/main/java/io/ballerina/stdlib/ftp/transport/RemoteFileSystemConnectorFactory.java
@@ -21,8 +21,10 @@ package io.ballerina.stdlib.ftp.transport;
 import io.ballerina.stdlib.ftp.exception.RemoteFileSystemConnectorException;
 import io.ballerina.stdlib.ftp.transport.client.connector.contract.VfsClientConnector;
 import io.ballerina.stdlib.ftp.transport.listener.RemoteFileSystemListener;
+import io.ballerina.stdlib.ftp.transport.server.FileDependencyCondition;
 import io.ballerina.stdlib.ftp.transport.server.connector.contract.RemoteFileSystemServerConnector;
 
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -31,7 +33,7 @@ import java.util.Map;
 public interface RemoteFileSystemConnectorFactory {
 
     /**
-     * @param connectorConfig    properties required for the {@link RemoteFileSystemServerConnector}.
+     * @param connectorConfig          properties required for the {@link RemoteFileSystemServerConnector}.
      * @param remoteFileSystemListener listener which gets triggered when message comes.
      * @return RemoteFileSystemServerConnector RemoteFileSystemServerConnector instance.
      * @throws RemoteFileSystemConnectorException if any error occurred when creating the server connector.
@@ -42,6 +44,17 @@ public interface RemoteFileSystemConnectorFactory {
 
     /**
      * @param connectorConfig          properties required for the {@link VfsClientConnector}.
+     * @param remoteFileSystemListener listener which gets triggered when message comes.
+     * @return VFSClientConnector instance
+     * @throws RemoteFileSystemConnectorException if any error occurred when initializing the server connector.
+     */
+    RemoteFileSystemServerConnector createServerConnector(Map<String, String> connectorConfig,
+                                                          List<FileDependencyCondition> dependencyConditions,
+                                                          RemoteFileSystemListener remoteFileSystemListener)
+            throws RemoteFileSystemConnectorException;
+
+    /**
+     * @param connectorConfig properties required for the {@link VfsClientConnector}.
      * @return VFSClientConnector instance
      * @throws RemoteFileSystemConnectorException if any error occurred when initializing the server connector.
      */