+++ native/src/main/java/io/ballerina/stdlib/ftp/transport/server/connector/contract/RemoteFileSystemServerConnector.java
@@ -47,4 +47,5 @@ public interface RemoteFileSystemServerConnector {
      *
      */
     FtpListener getFtpListener();
+
 }